router and main layout
